refactor(t1487): split email-to-markdown.py into focused modules
Reduces file-complexity from 223 to ~43 by extracting three pipeline modules:
- email_parser.py: MIME parsing, headers, body, attachment extraction (~274 lines, complexity ~48)
- email_normaliser.py: section normalisation, thread reconstruction, frontmatter (~455 lines, complexity ~68)
- email_md_summary.py: LLM/heuristic summary generation (~259 lines, complexity ~35)
- email-to-markdown.py: pipeline orchestrator + CLI only (~448 lines, complexity ~43)

Public API (email_to_markdown() 9-param signature) and CLI interface unchanged.
All modules pass syntax check. CLI --help verified working.

Closes #4907
